+++ EscapeFromEnceladus/EscapeFromEnceladus/Weapon/Missile.cs
@@ -49,7 +49,7 @@ namespace Enceladus.Weapon {
             get { return Flags; }
         }
 
-        public override int BaseDamage {
+        public override float BaseDamage {
             get { return 5; }
         }
     }
